requirements.txt
python-dotenv
langchain-openai==0.2.0
langchain==0.3.0
streamlit